Make filter be type of Vec<u8>
This commit change filter to type of Vec<u8> and also refactor a bit to
make the module more idiomatic.

Note that most parameters that need a Vector(Vec<u8>) can take a
slice(&[u8]) instead. Because Vector has a Deref trait which can
dereference to slice and rust will auto convert it when passing
arguments to functions. This is similar to the relationship between
String and &str.